+++ src/assault/game/gameObjects/Selectable.java
@@ -19,7 +19,7 @@ public class Selectable extends AObject implements MouseListener {
 
 	private boolean selected = false;
 
-	public Selectable(GameArea g, double x, double y, SelectableResourceHolder src, int health, Player owner) throws ResourceException {
+	public Selectable(GameArea g, float x, float y, SelectableResourceHolder src, int health, Player owner) throws ResourceException {
 		super(g, x, y, src, owner);
 	}
