+++ src/assault/game/loading/resourceHolders/ResourceResourceHolder.java
@@ -26,7 +26,7 @@ public class ResourceResourceHolder extends SelectableResourceHolder {
 	}
 
 	@Override
-	public Resource createObject(GameArea g, double x, double y, Player owner) {
+	public Resource createObject(GameArea g, float x, float y, Player owner) {
 		try {
 			return new Resource(g, x, y, this);
 		} catch (ResourceException ex) {
